Move application files into own module
