+++ README.md
@@ -8,8 +8,17 @@ pip install flask-restful
 pip install flask-cors
 ```
 
+## Creating the Database
+
+```
+$ python
+>>> from src.app import app, db
+>>> with app.app_context():
+...     db.init_app(app)
+```
+
 ## Running
 
 ```
-python app.py
+python main.py
 ```
