+++ src/app.py
@@ -2,11 +2,11 @@ from flask import Flask
 from flask_cors import CORS
 from flask_restful import Api
 
-from resources.twitter import Twitter
-from resources.attribute import Attribute
-from resources.playlist import Playlist
+from src.resources.twitter import Twitter
+from src.resources.attribute import Attribute
+from src.resources.playlist import Playlist
 
-from models.recommendation import Recommendation, db
+from src.models.recommendation import Recommendation, db
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///test.db'
