+++ src/models/shared.py
