+++ test/bob_tests/test_bob_twitter.py
@@ -1,4 +1,4 @@
-from songbird_server.resources.twitter import Twitter
+from src.resources.twitter import Twitter
 
 data = {
     "JeffBezos": 1,
